Moved DO configuration to new folder.
